+ fixed ToolbarActionManager titles + fixed adding/editing entities + enforced max number of hours per day + caches now return view model lists ordered by id, starting with newest + view models are now cloneable + getMatching method in AsyncGenericDao.java now takes an optional long that is used as parent id AsyncTimeEntriesDao.java: + uses a different algorithm to use the string search term (the string needs to be composed out of 3 longs (start date, end date, user id) separated by spaces ('temporary' solution for getting matching data out of the db) + customized error messages respective to their use cases ~ fixed the delete method + ActionListeners now contain the entity id needed by the ToolbarActionManager to create correct menus and actions TimeEntryAddFormFragment.java, TimeEntryEditFormFragment.java: + fixed some problems with validation, form data fill, entity saving + HoursPerDayValidation.java - validation provider for time entries max hours per user + added project name in time entry edit/add form KNOWN ISSUES: + the (recycler view) lists of items (view models) are not shown correctly when not stretched over whole vertical surface and more than 3-4 items are added + if a project/issue/time entry is modified, but not saved, the change is saved in cache (but not in the db)

+++ app/src/main/java/org/rares/miner49er/domain/issues/adapter/IssuesAdapter.java
@@ -141,7 +141,7 @@ public class IssuesAdapter extends AbstractAdapter<IssuesViewHolder> {
                 if (position == sp) {
                     if (payload instanceof Bundle) {
                         Bundle p = (Bundle) payload;
-                        int id = p.getInt("id");
+                        long id = p.getLong("id");
                         int color = p.getInt("Color", 0);
                         Log.d(TAG, "onChanged: >>>>> TRIGGERED >> new color >> " + color);
                         if (color != 0) {
@@ -161,6 +161,10 @@ public class IssuesAdapter extends AbstractAdapter<IssuesViewHolder> {
     @Override
     public void accept(List list) throws Exception {
 //        Log.d(TAG, "accept() called with: list = [" + list + "]");
+//        for (int i = 0; i < list.size(); i++) {
+//            IssueData issueData = (IssueData) list.get(i);
+//            Log.d(TAG, "accept: " + issueData.getName());
+//        }
         updateData(list);
     }
 }
